ai-services/requirements.txt
fastapi==0.118.0
google-cloud-storage==3.4.1
inference-sdk==0.58.2
langgraph==0.6.8
langchain-google-genai==2.1.12
pydantic==2.11.9
pydantic_settings==2.11.0
